backend/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.1
django-allauth==0.41.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.8
Markdown==3.1.1
oauthlib==3.1.0
Pillow==6.2.1
psycopg2==2.8.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
